requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
inotify==0.2.10
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==2.1.5
nose==1.3.7
Werkzeug==3.0.4
